+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fixed an issue where the **generate-test-playbook** command used an incorrect argument name as default
 * Fixed an issue where the **json-to-outputs** command used an incorrect argument name as default when using `-d`.
 * Fixed an issue where validations failed while trying to validate non content files.
+* Fixed an issue where README validations did not work post VS Code formatting.
 
 # 1.4.6
 * Fixed an issue where **validate** suggests, with no reason, running **format** on missing mandatory keys in yml file.