+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
+
+# 1.4.7
 * Fixed an issue where the **format** command has incorrectly recognized on which files to run when running using git.
 * Fixed an issue where author image validations were not checked properly.
 * Fixed an issue where new old-formatted scripts and integrations were not validated.